requirements.txt
flask==2.2.2
Pillow-SIMD==9.0.0post1
humanize==4.3.0